Separate `harbor.py` into modules and adapt `nuitka` (#53)
* spread modules

* nuitka mode change

* fix imports and successful nuitka build

* extract file check

* convert to asynchronous functions

* imports in separated files

* util functions

* harbor main file doc

* linter linebreaks and newlines

* remove trailing whitespace

* fix missing import

* fix missing newline

* nuitka onefile option

* remove global client

* fix line linter

* add missing import

* extract path

* fix missing and unneeded imports

* fix missing and unneeded imports of separated files

* changed positions of env vars

* updated import

* moved enum

* enum import to file

* rename variable to new name

* add new_admin_password to utils
